+++ lib/flexi_kline.dart
@@ -19,8 +19,7 @@ export 'src/constant.dart';
 export 'src/core/core.dart';
 export 'src/data/kline_data.dart';
 export 'src/extension/export.dart';
-export 'src/framework/export.dart'
-    hide PaintDelegateExt, MainPaintDelegateExt, MainPaintManagerExt, IConfigurationExt, ILogger, FlexiLog;
+export 'src/framework/export.dart' hide PaintDelegateExt, MainPaintDelegateExt, MainPaintManagerExt, IConfigurationExt;
 export 'src/indicators/export.dart';
 export 'src/kline_controller.dart';
 export 'src/model/export.dart';